requirements.txt
django
gunicorn
django-heroku
python-dateutil
dload
Pillow
pyzbar

